Add circle's state and Threshold Circle
This commit adds the logic to keep each circle's states per
each location in the map, defined as the polygons/union of polygons.

The kml is processed and each location is stored in a map object,
calculating and storing 4 circles per location:
- large, small, custom and threshold

Custom circle settings are recalculated on each slider modification,
but the others are calculated when parsing the KML file and stored in
the locations map object, instead of being calculated on each location
selection change in the "Select Targeting Area" selector.

By default, the threshold circle is selected to be downloaded. This can
be modified with the new "Selected Settings" radio buttons

The csv downloaded will download the selected circles per location,
being able to mix Custom Circles, Threshold Circles, etc

Change-Id: I2f86e59de191e3c3f3a8a3b7c61d27837ef193a3